+++ packages/all/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @typeschema/all
 
+## 0.13.8
+
+### Patch Changes
+
+- a7c09b6: feat: add support to fastest-validator
+- Updated dependencies [a7c09b6]
+  - @typeschema/fastest-validator@0.1.0
+  - @typeschema/main@0.13.8
+
 ## 0.13.7
 
 ### Patch Changes